Update Entity ID type and add logic for subscribe and unsubscribe methods

+++ subscription-service/src/main/java/com/adidas/subscriptionservice/repository/ISubscriptionRepository.java
@@ -3,10 +3,8 @@ package com.adidas.subscriptionservice.repository;
 import com.adidas.subscriptionservice.entity.Subscription;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.math.BigInteger;
-
 public interface ISubscriptionRepository extends JpaRepository<Subscription, Long> {
 
-    Subscription findById(BigInteger id);
+    Subscription findById(Integer id);
 
 }
